revamp homepage and move all products to a different section
