+++ frontend/src/Pages/Home.jsx
@@ -1,112 +1,56 @@
-import { Suspense, lazy, useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import useAxiosSecure from "../hooks/useAxiosSecure";
-const AllProductCard = lazy(() =>
-    import("../components/shared/AllProductCard")
-);
-const LoadingCard = lazy(() => import("../components/shared/LoadingCard"));
-import SearchBar from "../components/shared/SearchBar";
-import Loader from "../components/shared/Loader";
+import { AuthContext } from "../Auth/AuthProvider";
+import ProductCard from "../components/shared/ProductCard";
 
 const Home = () => {
-    const [allProducts, setAllProducts] = useState([]);
-    const [sortedProducts, setSortedProducts] = useState([]);
-    const [loading, setLoading] = useState(false);
-    const [sortOption, setSortOption] = useState("");
     const axiosSecure = useAxiosSecure();
+    const { loading, setLoading } = useContext(AuthContext);
+    const [bestSellers, setBestSellers] = useState([]);
 
     useEffect(() => {
         setLoading(true);
         axiosSecure
-            .get(
-                "/products?attributes=Product_ID,Name,Price,Galaxy_source,Planet_source,Quantity_inStock,Image_Url"
-            )
-            .then((response) => {
-                setAllProducts(response.data);
+            .get("/bestSellers")
+            .then((res) => {
+                setBestSellers(res.data);
+                setLoading(false);
             })
-            .catch((error) => {
-                console.error(error);
-            })
-            .finally(() => {
+            .catch((err) => {
+                console.log(err);
                 setLoading(false);
             });
     }, []);
 
-    useEffect(() => {
-        if (sortOption) {
-            const sorted = [...allProducts].sort((a, b) => {
-                switch (sortOption) {
-                    case "name":
-                        return a.Name.localeCompare(b.Name);
-                    case "galaxy":
-                        return a.Galaxy_source.localeCompare(b.Galaxy_source);
-                    case "planet":
-                        return a.Planet_source.localeCompare(b.Planet_source);
-                    case "price":
-                        return a.Price - b.Price;
-                    default:
-                        return 0;
-                }
-            });
-            setSortedProducts(sorted);
-        } else {
-            setSortedProducts(allProducts);
-        }
-    }, [sortOption, allProducts]);
-
-    if (loading) {
-        return (
-            <div className="mt-24 mx-3 lg:mx-auto text-white lg:max-w-7xl">
-                <div className="grid lg:grid-cols-3 grid-cols-1 gap-4">
-                    <Suspense fallback={<Loader/>}>
-                        <LoadingCard />
-                        <LoadingCard />
-                        <LoadingCard />
-                    </Suspense>
-                </div>
-            </div>
-        );
-
-    }
-
     return (
-        <div className="mt-24 mx-3 lg:mx-auto text-white lg:max-w-7xl">
-            <div className="flex justify-center items-center mb-4 lg:flex-row flex-col gap-4">
-                <SearchBar
-                    setAllProducts={setAllProducts}
-                    setLoading={setLoading}
-                />
-                <div>
-                    <label htmlFor="sort" className="mr-2">
-                        Sort by:
-                    </label>
-                    <select
-                        id="sort"
-                        value={sortOption}
-                        onChange={(e) => setSortOption(e.target.value)}
-                        className="bg-primary text-white p-2 rounded-3xl space-y-2"
-                    >
-                        <option value="">Select</option>
-                        <option value="name">Name</option>
-                        <option value="galaxy">Galaxy</option>
-                        <option value="planet">Planet</option>
-                        <option value="price">Price</option>
-                    </select>
+        <div className="mt-[80px] px-3 lg:mx-auto text-white glass pb-10">
+            <div className="flex lg:flex-row flex-col-reverse items-center justify-center">
+                <div className="lg:max-w-3xl flex flex-col justify-end items-end ml-16">
+                    <h1 className="text-7xl font-bold bg-gradient-to-r from-tertiary via-secondary to-primary text-transparent bg-clip-text animate-gradient bg-300%">
+                        Discover the Galaxy&apos;s Best Products
+                    </h1>
+                    <p className="text-3xl font-semibold text-gray-400">
+                        Explore our vast collection of intergalactic goods, from
+                        cutting-edge technology to rare alien artifacts.
+                    </p>
                 </div>
+                <img
+                    src="https://i.ibb.co/sRP1gWS/OIG1.jpg"
+                    alt=""
+                    className="max-h-[calc(100dvh-80px)] lg:rounded-bl-lg"
+                />
             </div>
-            {!loading && (
-                <div className="grid lg:grid-cols-3 grid-cols-1 gap-4">
-                    {sortedProducts.map((product) => (
-                            <AllProductCard product={product} key={product.Product_ID} />
+            <div className="mt-16 lg:max-w-7xl lg:mx-auto mx-3">
+                <h2 className="text-4xl font-bold text-center underline underline-offset-4">Best Sellers</h2>
+                <div className="flex flex-wrap justify-center gap-5 mt-8">
+                    {bestSellers.map((product) => (
+                        <ProductCard
+                            key={product.Product_ID}
+                            product={product}
+                        />
                     ))}
                 </div>
-            )}
-            {allProducts.length === 0 && !loading && (
-                <div className="flex justify-center items-center h-[70vh]">
-                    <h1 className="text-3xl text-white bg-black px-3 py-2 rounded-lg">
-                        No Products Found
-                    </h1>
-                </div>
-            )}
+            </div>
         </div>
     );
 };
